Initialise VQE in gst_vqesrc_class_init rather than start
We need to ensure that VQE-C globals are initialised only once per
process otherwise it is an error so this patch moves initialisation
from state change time to class initialisation time.  This seems to
fit in some ways as classes are static information common to all
objects of a type as are the statics in VQE.  Nonetheless this is
really ugly in other ways as initialising VQE actually involves
reading a configuration file and the parameters it sets aren't fixed
at compile time, unlike much of the information in GstVQESrcClass.

This does also have the positive effect that you no-longer see lots
of errors when calling `gst-inspect-1.0 vqesrc` associated with
collecting statistics.