gstvqesdpdemux:  Name the VQE element with "vqesrc"
This is to make it easier to find the element in the pipeline by name when
compared to rather ambiguous "src".

It might be considered a sort of API for a GstBin element to expose its
child elements with well known names (if a GstBin has "static" children
list, it's displayed by `gst-inspect-1.0' in the "Children" section).  It
makes referring to those elements from the pipeline (application) level
much easier.