Add a property that would flag that VQE stats are valid
We need a way to know if the VQE stats are valid before we start reading
them. This is because it can be quite expensive to receive 40 odd async
error messages. Instead a new property has been added that is true if
reading any of the VQE stats property would be error free.